Remove visitor.hpp, replace more pointer to references
