+++ include/frontend/ast/ast_container.hpp
@@ -27,11 +27,11 @@ private:
   std::vector<i_ast_node_uptr> m_nodes;
   i_ast_node                  *m_root = nullptr;
 
-  template <typename T, typename... Ts> T *emplace_back(Ts &&...args) {
+  template <typename T, typename... Ts> T &emplace_back(Ts &&...args) {
     auto uptr = std::make_unique<T>(std::forward<Ts>(args)...);
     auto ptr = uptr.get();
     m_nodes.push_back(std::move(uptr));
-    return ptr;
+    return *ptr;
   }
 
 public:
@@ -62,7 +62,7 @@ public:
   i_ast_node *get_root_ptr() const { return m_root; }
 
   template <typename t_node_type, typename... t_args>
-  t_node_type *make_node(t_args &&...args)
+  t_node_type &make_node(t_args &&...args)
   requires std::is_base_of_v<i_ast_node, t_node_type>
   {
     return emplace_back<t_node_type>(std::forward<t_args>(args)...);
