+++ include/frontend/ast/ast_nodes.hpp
@@ -22,4 +22,3 @@
 #include "ast_nodes/unary_expression.hpp"
 #include "ast_nodes/variable_expression.hpp"
 #include "ast_nodes/while_statement.hpp"
-#include "visitor.hpp"
