+++ include/frontend/ast/ast_nodes/i_ast_node.hpp
@@ -28,4 +28,21 @@ public:
   virtual ~i_ast_node() {}
 };
 
+class i_ast_node;
+class assignment_statement;
+class binary_expression;
+class constant_expression;
+class if_statement;
+class print_statement;
+class read_expression;
+class statement_block;
+class unary_expression;
+class variable_expression;
+class while_statement;
+class error_node;
+
+using tuple_ast_nodes =
+    std::tuple<assignment_statement, binary_expression, constant_expression, if_statement, print_statement,
+               read_expression, statement_block, unary_expression, variable_expression, while_statement, error_node>;
+
 } // namespace paracl::frontend::ast
