+++ src/frontend/parser.y
@@ -31,7 +31,6 @@
 #include "frontend/ast/ast_container.hpp"
 #include "frontend/ast/ast_copier.hpp"
 #include "frontend/ast/ast_nodes.hpp"
-#include "frontend/ast/visitor.hpp"
 
 namespace paracl::frontend {
   class scanner;
